api/requirements.txt
fastapi==0.66.1
uvicorn==0.14.0
python-dotenv==0.18.0
scikit-learn==0.24.2
pymongo==3.12.0
dnspython==2.0.0
pika==1.2.0